fix(voice): average token embeddings without integer division
out[i] /= tokens.size performed integer division when token count was not a
multiple of the embedding dimension's divisor, skewing the pooled vector and
mangling phrase similarity. Cast to Float before dividing.

Also ship the SOME/IP native bridge (libmotorguardsomeip.so) and libc++ so
Gradle aaos builds carry the diagnostics link alongside the AOSP drop-in's
deploy-time patch.

+++ MotorGuardApp/app/src/main/java/com/motorguard/ivi/ui/voice/nlu/TextEmbedder.kt
@@ -72,7 +72,7 @@ class TextEmbedder private constructor(
         val dim = tokens.firstOrNull()?.size ?: return FloatArray(0)
         val out = FloatArray(dim)
         for (token in tokens) for (i in 0 until dim) out[i] += token[i]
-        for (i in 0 until dim) out[i] /= tokens.size
+        for (i in 0 until dim) out[i] /= tokens.size.toFloat()
         return out
     }
